move atom to internal folder
